+++ Cache/RedisCache.cs
@@ -44,6 +44,8 @@ namespace CYQ.Data.Cache
 
         public override bool Set(string key, object value, double cacheMinutes, string fileName)
         {
+            if (string.IsNullOrEmpty(key)) { return false; }
+            if (value == null) { return Remove(key); }
             return client.Set(key, value, Convert.ToInt32(cacheMinutes * 60));
         }
 
@@ -99,6 +101,7 @@ namespace CYQ.Data.Cache
 
         public override bool Contains(string key)
         {
+            if (string.IsNullOrEmpty(key)) { return false; }
             return client.ContainsKey(key);
         }
 
@@ -139,12 +142,14 @@ namespace CYQ.Data.Cache
 
         public override object Get(string key)
         {
+            if (string.IsNullOrEmpty(key)) { return null; }
             return client.Get(key);
         }
 
 
         public override bool Remove(string key)
         {
+            if (string.IsNullOrEmpty(key)) { return false; }
             return client.Delete(key);
         }
